+++ src/main/kotlin/io/runescript/plugin/lang/psi/RsElementGenerator.kt
@@ -1,11 +1,16 @@
 package io.runescript.plugin.lang.psi
 
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.util.text.StringUtil
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiFileFactory
+import com.intellij.psi.PsiWhiteSpace
+import com.intellij.psi.TokenType
 import com.intellij.psi.impl.PsiFileFactoryImpl
+import com.intellij.psi.tree.TokenSet
 import com.intellij.psi.util.PsiTreeUtil
+import com.intellij.psi.util.childrenOfType
 import com.intellij.testFramework.LightVirtualFile
 import io.runescript.plugin.ide.filetypes.RsFileType
 import io.runescript.plugin.lang.RuneScript
@@ -17,15 +22,18 @@ object RsElementGenerator {
         return PsiTreeUtil.findChildOfType(element, RsExpression::class.java) as RsExpression
 
     }
+
     fun createIntegerLiteral(project: Project, replacement: String): PsiElement {
         val element = createDummyFile(project, "[proc,dummy]()()$replacement;")
-        val literal = PsiTreeUtil.findChildOfType(element, RsIntegerLiteralExpression::class.java) as RsIntegerLiteralExpression
+        val literal =
+            PsiTreeUtil.findChildOfType(element, RsIntegerLiteralExpression::class.java) as RsIntegerLiteralExpression
         return literal.node.firstChildNode.psi!!
     }
 
     fun createColorTag(project: Project, color: Int, tagName: String = "col"): PsiElement {
         val element = createDummyFile(project, "[proc,dummy]()()\"<$tagName=%06x>\";".format(color and 0xffffff))
-        val literal = PsiTreeUtil.findChildOfType(element, RsStringLiteralExpression::class.java) as RsStringLiteralExpression
+        val literal =
+            PsiTreeUtil.findChildOfType(element, RsStringLiteralExpression::class.java) as RsStringLiteralExpression
         return literal.node.findChildByType(RsElementTypes.STRING_TAG)!!.psi
     }
 
@@ -41,14 +49,19 @@ object RsElementGenerator {
 
     fun createStringLiteralContent(project: Project, content: String): RsStringLiteralContent {
         val element = createDummyFile(project, "[proc,dummy]()()\"$content\";")
-        val literal = PsiTreeUtil.findChildOfType(element, RsStringLiteralExpression::class.java) as RsStringLiteralExpression
+        val literal =
+            PsiTreeUtil.findChildOfType(element, RsStringLiteralExpression::class.java) as RsStringLiteralExpression
         return PsiTreeUtil.findChildOfType(literal, RsStringLiteralContent::class.java) as RsStringLiteralContent
     }
 
     private fun createDummyFile(project: Project, text: String): PsiFile {
         val factory = PsiFileFactory.getInstance(project) as PsiFileFactoryImpl
         val name = "dummy.${RsFileType.defaultExtension}"
-        val virtualFile = LightVirtualFile(name, RsFileType, text)
+        val virtualFile = LightVirtualFile(name, RsFileType, StringUtil.convertLineSeparators(text))
         return factory.trySetupPsiForFile(virtualFile, RuneScript, false, true)!!
+    }
+
+    fun createNewLine(project: Project): PsiElement {
+        return createDummyFile(project, "\n").firstChild
     }
 }